feat: add style and meaning to information section
Added unique images related to the a hypothetical gym landing page. One
of the images is not related. Can you guess which one? Also updated
placeholder text in the information section. Lmao I don't know if this
is optimal for a landing page. Need to learn sales and marketing better.